requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
googletrans==3.1.0a0
python-dotenv==0.19.0
gunicorn==20.1.0
deep-translator==1.11.4
PyPDF2==3.0.1
reportlab==4.0.8
arabic-reshaper==3.0.0
python-bidi==0.4.2
